Refactor imports to move lookup from libs.data to libs.functions for better organization
