+++ integrations/slack/events/comparison.py
@@ -6,9 +6,10 @@ import logging
 from typing import TYPE_CHECKING, cast
 
 import libs.global_value as g
-from libs.data import lookup, modify, search
+from libs.data import modify, search
 from libs.domain.datamodels import ComparisonResults
 from libs.domain.score import GameResult
+from libs.functions import lookup
 from libs.types import ActionStatus, CommandType, StyleOptions
 from libs.utils import formatter, validator
 from libs.utils.timekit import ExtendedDatetime as ExtDt
