+++ integrations/slack/functions.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any, Union, cast
 
 import libs.global_value as g
 from integrations.base.interface import FunctionsInterface
-from libs.data import lookup
+from libs.functions import lookup
 from libs.types import ActionStatus
 from libs.utils import validator
 from libs.utils.timekit import Delimiter, Format
