+++ libs/data/__init__.py
@@ -2,7 +2,6 @@
 データ操作モジュール
 
 - `libs.data.aggregate`: データ集計
-- `libs.data.lookup`: 情報取得
 - `libs.data.modify`: DBデータ操作
 - `libs.data.search`: DB検索
 """
