+++ libs/functions/lookup.py
@@ -1,5 +1,5 @@
 """
-libs/data/lookup.py
+libs/functions/lookup.py
 """
 
 import logging
