+++ libs/functions/tools/unification.py
@@ -6,7 +6,8 @@ import configparser
 import logging
 
 import libs.global_value as g
-from libs.data import lookup, modify
+from libs.data import modify
+from libs.functions import lookup
 from libs.utils import dbutil, textutil, validator
