+++ tests/database/conftest.py
@@ -14,7 +14,7 @@ import pytest
 import libs.global_value as g
 from libs.bootstrap import configuration, initialization
 from libs.bootstrap.app_config import AppConfig
-from libs.data import lookup
+from libs.functions import lookup
 from libs.utils import dbutil
 
 if TYPE_CHECKING:
